@@ -5,7 +5,7 @@
   "main": "./src/naoko.ts",
   "scripts": {
     "test": "ts-mocha -p ./tsconfig.json ./src/tests/logic.test.ts",
-    "start": "ts-node -T ./src/naoko.ts",
+    "start": "node --experimental-specifier-resolution=node --loader ts-node/esm ./src/naoko.ts",
     "dev": "nodemon ./src/naoko.ts",
     "setup": "ts-node ./src/setup.naoko.ts",
     "first-time-setup": "npm ci && npm run prettify && ts-node ./src/setup.naoko.ts",
@@ -27,36 +27,37 @@
     "url": "https://github.com/Geoxor/naoko/issues"
   },
   "homepage": "https://github.com/Geoxor/naoko#readme",
+  "type": "module",
   "dependencies": {
-    "@discordjs/opus": "^0.5.3",
-    "@discordjs/voice": "^0.5.5",
+    "@discordjs/opus": "^0.9.0",
+    "@discordjs/voice": "^0.16.0",
     "@hbis/three-obj-mtl-loader": "^1.1.0",
-    "@jimp/plugin-fisheye": "^0.16.1",
+    "@jimp/plugin-fisheye": "^0.22.8",
     "@taku.moe/pura": "^1.0.4",
     "@taku.moe/utils": "^1.0.9",
-    "axios": "^0.21.1",
+    "axios": "^1.4.0",
     "canvas": "^2.8.0",
-    "chalk": "^4.1.1",
-    "discord.js": "^13.4.0",
-    "file-type": "^9.0.0",
-    "get-image-colors": "^2.0.1",
+    "chalk": "^5.2.0",
+    "discord.js": "^14.11.0",
+    "file-type": "^18.4.0",
+    "get-image-colors": "^4.0.1",
     "gifenc": "^1.0.3",
-    "jimp": "^0.16.1",
+    "jimp": "^0.22.8",
     "js-levenshtein": "^1.1.6",
-    "mongoose": "^6.0.12",
+    "mongoose": "^7.2.1",
     "multi-progress": "^4.0.0",
-    "music-metadata": "^7.11.3",
-    "node": "16.6.1",
-    "node-canvas-webgl": "^0.2.6",
+    "music-metadata": "^8.1.4",
+    "node": "20.2.0",
+    "node-canvas-webgl": "^0.2.7",
     "pet-pet-gif": "^1.0.9",
     "progress": "^2.0.3",
     "rcon.js": "^1.0.0",
     "replace-last": "^1.2.6",
     "systeminformation": "^5.8.0",
-    "three": "^0.131.3",
+    "three": "^0.131.1",
     "three-obj-mtl-loader": "^1.0.3",
     "tweetnacl": "^1.0.3",
-    "typescript": "^4.3.5"
+    "typescript": "^5.0.4"
   },
   "devDependencies": {
     "@types/chai": "^4.2.21",
@@ -64,12 +65,12 @@
     "@types/get-image-colors": "^4.0.1",
     "@types/gifencoder": "^2.0.1",
     "@types/js-levenshtein": "^1.1.0",
-    "@types/mocha": "^9.0.0",
+    "@types/mocha": "^10.0.1",
     "@types/multi-progress": "^2.0.3",
     "@types/mv": "^2.1.2",
     "@types/three": "^0.131.0",
     "chai": "^4.3.4",
-    "mocha": "^9.0.3",
+    "mocha": "^10.2.0",
     "mv": "^2.1.1",
     "nodemon": "^2.0.12",
     "ts-node": "^10.4.0"
