@@ -7,7 +7,7 @@ export default defineCommand({
   aliases: ["cls"],
   usage: "clear <amount>",
   description: "Bulk delete messages up to 100",
-  permissions: ["MANAGE_MESSAGES"],
+  permissions: ["ManageMessages"],
   execute: async (message) => {
     let count = parseFloat(message.args[0]) + 1;
     count = count > 100 ? 100 : count;
