@@ -1,4 +1,4 @@
-import Discord from "discord.js";
+import { ChannelType } from "discord.js";
 import { GEOXOR_GUILD_ID, TESTING_GUILD_ID } from "../constants";
 import { isDiscordInvite } from "../moderation/isDiscordInvite";
 import { isFreeNitro } from "../moderation/isFreeNitro";
@@ -9,7 +9,7 @@ const checks = [isFreeNitro, isDiscordInvite];
 
 export default async function (message: IMessage, next: (message: IMessage) => any): Promise<void> {
   try {
-    if (!(message.guild?.id === GEOXOR_GUILD_ID || message.guild?.id === TESTING_GUILD_ID || message.channel.type == "DM"))
+    if (!(message.guild?.id === GEOXOR_GUILD_ID || message.guild?.id === TESTING_GUILD_ID || message.channel.type == ChannelType.DM))
       return;
     for (let i = 0; i < checks.length; i++) {
       const checkFn = checks[i];
