@@ -2,7 +2,7 @@ import Discord from "discord.js";
 import { randomChoice } from "../logic/logic";
 import { definePlugin } from "../naoko/Plugin";
 import { defineCommand } from "../types";
-import facts from "./fun-facts/facts.json";
+import facts from "./fun-facts/facts.json" assert { type: 'json' };
 
 export default definePlugin({
   name: "@azur1s/fun-fact",
@@ -13,9 +13,9 @@ export default definePlugin({
     usage: "funfact",
     description: "Random fun fact",
     execute: () => {
-      const embed = new Discord.MessageEmbed()
+      const embed = new Discord.EmbedBuilder()
         .setColor("#d2185e")
-        .setAuthor(`Fun Facts by azur`)
+        .setAuthor({ name: 'Fun Facts by azur' })
         .setThumbnail(randomChoice([
           'https://wiki.hypixel.net/images/0/0a/SkyBlock_items_enchanted_book_and_quill.gif',
           'https://wiki.hypixel.net/images/4/4e/SkyBlock_items_enchanted_book.gif'
@@ -25,7 +25,7 @@ export default definePlugin({
           value: `${randomChoice(facts)}`,
           inline: true,
         })
-        .setFooter(`Total of ${facts.length} fun facts`);
+        .setFooter({ text: `Total of ${facts.length} fun facts` });
 
       return { embeds: [embed] };
     },
